Added resource links to prerequisites section
Since we're telling readers that they should be familiar with these concepts, it's probably a good idea to link them to resources they can use to learn about the concepts, in case they're not already familiar with them. These may not each be the absolute best resource available, but in all except one case I linked to the official PHP documentation, so it's (hopefully) a good starting point.